flask-application/requirements.txt
Flask
mysql-connector-python
qrcode
pyqrcode
gunicorn

